+++ src/lib.rs
@@ -24,7 +24,6 @@ pub use self::r#const::*;
 pub use self::vyper::combined_json::contract::Contract as VyperCompilerCombinedJsonContract;
 pub use self::vyper::combined_json::CombinedJson as VyperCompilerCombinedJson;
 pub use self::vyper::standard_json::input::language::Language as VyperCompilerStandardInputJsonLanguage;
-pub use self::vyper::standard_json::input::settings::evm_version::EVMVersion as VyperCompilerStandardInputJsonSettingsEVMVersion;
 pub use self::vyper::standard_json::input::settings::selection::Selection as VyperCompilerStandardInputJsonSettingsSelection;
 pub use self::vyper::standard_json::input::settings::Settings as VyperCompilerStandardInputJsonSettings;
 pub use self::vyper::standard_json::input::source::Source as VyperCompilerStandardInputJsonSource;
@@ -108,9 +107,11 @@ pub fn zkasm(
 ///
 /// Runs the standard output mode.
 ///
+#[allow(clippy::too_many_arguments)]
 pub fn standard_output(
     input_files: Vec<PathBuf>,
     vyper: &VyperCompiler,
+    evm_version: Option<compiler_llvm_context::EVMVersion>,
     vyper_optimizer_enabled: bool,
     optimizer_settings: compiler_llvm_context::OptimizerSettings,
     include_metadata_hash: bool,
@@ -119,12 +120,17 @@ pub fn standard_output(
 ) -> anyhow::Result<Build> {
     if let Some(ref debug_config) = debug_config {
         for path in input_files.iter() {
-            let lll_debug = vyper.lll_debug(path.as_path(), true)?;
+            let lll_debug = vyper.lll_debug(path.as_path(), evm_version, true)?;
             debug_config.dump_lll(path.to_string_lossy().as_ref(), lll_debug.as_str())?;
         }
     }
 
-    let project = vyper.batch(&vyper.version.default, input_files, vyper_optimizer_enabled)?;
+    let project = vyper.batch(
+        &vyper.version.default,
+        input_files,
+        evm_version,
+        vyper_optimizer_enabled,
+    )?;
 
     let build = project.compile(
         optimizer_settings,
@@ -144,6 +150,7 @@ pub fn standard_output(
 pub fn combined_json(
     input_files: Vec<PathBuf>,
     vyper: &VyperCompiler,
+    evm_version: Option<compiler_llvm_context::EVMVersion>,
     vyper_optimizer_enabled: bool,
     optimizer_settings: compiler_llvm_context::OptimizerSettings,
     include_metadata_hash: bool,
@@ -156,7 +163,8 @@ pub fn combined_json(
 
     if let Some(ref debug_config) = debug_config {
         for path in input_files.iter() {
-            let lll_debug = vyper.lll_debug(path.as_path(), vyper_optimizer_enabled)?;
+            let lll_debug =
+                vyper.lll_debug(path.as_path(), evm_version, vyper_optimizer_enabled)?;
             debug_config.dump_lll(path.to_string_lossy().as_ref(), lll_debug.as_str())?;
         }
     }
@@ -164,6 +172,7 @@ pub fn combined_json(
     let project: Project = vyper.batch(
         &vyper.version.default,
         input_files.clone(),
+        evm_version,
         vyper_optimizer_enabled,
     )?;
 
@@ -175,7 +184,7 @@ pub fn combined_json(
         debug_config,
     )?;
 
-    let mut combined_json = vyper.combined_json(input_files.as_slice())?;
+    let mut combined_json = vyper.combined_json(input_files.as_slice(), evm_version)?;
     build.write_to_combined_json(&mut combined_json, &zkvyper_version)?;
 
     match output_directory {