+++ src/vyper/standard_json/input/mod.rs
@@ -44,6 +44,7 @@ impl Input {
         output_selection: BTreeMap<String, Vec<Selection>>,
         optimize: bool,
         fallback_to_optimizing_for_size: bool,
+        disable_system_request_memoization: bool,
     ) -> anyhow::Result<Self> {
         let sources = paths
             .into_par_iter()
@@ -63,6 +64,7 @@ impl Input {
                 output_selection,
                 optimize,
                 fallback_to_optimizing_for_size,
+                disable_system_request_memoization,
             ),
         })
     }
@@ -78,6 +80,7 @@ impl Input {
         output_selection: BTreeMap<String, Vec<Selection>>,
         optimize: bool,
         fallback_to_optimizing_for_size: bool,
+        disable_system_request_memoization: bool,
     ) -> anyhow::Result<Self> {
         let sources = sources
             .into_iter()
@@ -92,6 +95,7 @@ impl Input {
                 output_selection,
                 optimize,
                 fallback_to_optimizing_for_size,
+                disable_system_request_memoization,
             ),
         })
     }