+++ src/vyper/standard_json/input/settings/mod.rs
@@ -27,6 +27,9 @@ pub struct Settings {
     /// Whether to try to recompile with -Oz if the bytecode is too large.
     #[serde(skip_serializing)]
     pub fallback_to_optimizing_for_size: Option<bool>,
+    /// Whether to disable the system request memoization.
+    #[serde(skip_serializing)]
+    pub disable_system_request_memoization: Option<bool>,
 }
 
 impl Settings {
@@ -38,12 +41,14 @@ impl Settings {
         output_selection: BTreeMap<String, Vec<Selection>>,
         optimize: bool,
         fallback_to_optimizing_for_size: bool,
+        disable_system_request_memoization: bool,
     ) -> Self {
         Self {
             evm_version,
             output_selection,
             optimize,
             fallback_to_optimizing_for_size: Some(fallback_to_optimizing_for_size),
+            disable_system_request_memoization: Some(disable_system_request_memoization),
         }
     }
 }