+++ src/zkvyper/main.rs
@@ -55,7 +55,7 @@ fn main_inner() -> anyhow::Result<()> {
     }
 
     if arguments.recursive_process {
-        return compiler_vyper::run_process();
+        return era_compiler_vyper::run_process();
     }
 
     let debug_config = match arguments.debug_output_directory {
@@ -69,22 +69,24 @@ fn main_inner() -> anyhow::Result<()> {
     };
 
     let suppressed_warnings = match arguments.suppress_warnings {
-        Some(warnings) => compiler_vyper::WarningType::try_from_strings(warnings.as_slice())?,
+        Some(warnings) => era_compiler_vyper::WarningType::try_from_strings(warnings.as_slice())?,
         None => vec![],
     };
 
-    let vyper = compiler_vyper::VyperCompiler::new(
-        arguments
-            .vyper
-            .unwrap_or_else(|| compiler_vyper::VyperCompiler::DEFAULT_EXECUTABLE_NAME.to_owned()),
-    )?;
+    let vyper =
+        era_compiler_vyper::VyperCompiler::new(arguments.vyper.unwrap_or_else(|| {
+            era_compiler_vyper::VyperCompiler::DEFAULT_EXECUTABLE_NAME.to_owned()
+        }))?;
 
     let mut optimizer_settings = match arguments.optimization {
         Some(mode) => compiler_llvm_context::OptimizerSettings::try_from_cli(mode)?,
         None => compiler_llvm_context::OptimizerSettings::cycles(),
     };
     if arguments.fallback_to_optimizing_for_size {
-        optimizer_settings.set_fallback_to_size();
+        optimizer_settings.enable_fallback_to_size();
+    }
+    if arguments.disable_system_request_memoization {
+        optimizer_settings.disable_system_request_memoization();
     }
     optimizer_settings.is_verify_each_enabled = arguments.llvm_verify_each;
     optimizer_settings.is_debug_logging_enabled = arguments.llvm_debug_logging;
@@ -99,7 +101,7 @@ fn main_inner() -> anyhow::Result<()> {
     };
 
     let build = if arguments.llvm_ir {
-        compiler_vyper::llvm_ir(
+        era_compiler_vyper::llvm_ir(
             arguments.input_files,
             optimizer_settings,
             include_metadata_hash,
@@ -107,7 +109,7 @@ fn main_inner() -> anyhow::Result<()> {
             debug_config,
         )
     } else if arguments.zkasm {
-        compiler_vyper::zkasm(
+        era_compiler_vyper::zkasm(
             arguments.input_files,
             include_metadata_hash,
             suppressed_warnings,
@@ -116,7 +118,7 @@ fn main_inner() -> anyhow::Result<()> {
     } else {
         match arguments.format.as_deref() {
             Some("combined_json") => {
-                compiler_vyper::combined_json(
+                era_compiler_vyper::combined_json(
                     arguments.input_files,
                     &vyper,
                     !arguments.disable_vyper_optimizer,
@@ -132,7 +134,7 @@ fn main_inner() -> anyhow::Result<()> {
             Some(format) if format.split(',').any(|format| format == "combined_json") => {
                 anyhow::bail!("`combined_json` must be the only output format requested");
             }
-            Some(_) | None => compiler_vyper::standard_output(
+            Some(_) | None => era_compiler_vyper::standard_output(
                 arguments.input_files,
                 &vyper,
                 !arguments.disable_vyper_optimizer,