+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- The `Cancun` EVM version support with temporarily unsupported transient storage
+- The `MCOPY` instruction support
+- The `--evm-version` parameter to the CLI
+
 ## [1.3.17] - 2024-01-30
 
 ### Added