+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # The `zkvyper` changelog
 
+## [Unreleased]
+
+### Added
+
+- An option to disable the system request memoization
+
+## [1.3.17] - 2024-01-30
+
+### Added
+
+- More LLVM optimizations
+
 ## [1.3.16] - 2024-01-12
 
 ### Fixed